Only apply isort to known source files/dirs
For example, if you would create a virtual environment in the subfolder
venv, isort will check all of the installed packages in that venv.

This also fixes a small bug where the ">=5" at the end of the pip
install command for isort would be interpreted as a redirect, creating a
file "=5" with the output from pip.